backend/pyproject.toml
[tool.poetry]
name = "farm-docker"
version = "1.1.0"
description = "Backend API for farm-docker"
authors = ["Lawrence Linn <lawrence.rs.lin@gmail.com>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
pymongo = "^4.5.0"
fastapi = "^0.104.1"
uvicorn = "^0.23.2"
python-multipart = "^0.0.6"
motor = "^3.3.1"
pydantic = "^2.4.2"
email-validator = "^2.1.0.post1"
python-jose = "^3.3.0"
passlib = "^1.7.4"
bcrypt = "^4.0.1"
beanie = "^1.23.1"
python-dotenv = "^1.0.0"
fastapi-sso = "^0.7.2"
pydantic-settings = "^2.0.3"
websockets = "^10.0"
opencv-python = "4.8.1.78"
aiortc = "^1.3.0"
asyncio = "^3.4.3"
Pillow = "10.2.0"
torchvision = "0.17.0"
torch = "2.2.0"
#cuda = "12.1"
langchain = "0.1.13"
langchain-community = "0.0.29"
transformers = "4.39.1"
ollama = "0.1.7"
#bitsandbytes = "^0.0"
#pydantic-async = "0.1.0"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
black = "^23.10.1"
pre-commit = "^3.5.0"
httpx = "^0.25.0"
asgi-lifespan = "^2.1.0"
pytest = "^7.4.3"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
